requirements.txt
flask==3.0.2
flask-cors==4.0.0
transformers==4.39.3
torch==2.2.2
torchvision==0.17.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
opencv-python==4.9.0.80
PyMuPDF==1.23.26
Pillow==10.3.0
requests==2.31.0
